OCR/requirements.txt
Flask>=3.0.0,<4.0
PyMuPDF>=1.23.8,<1.26
pytesseract>=0.3.10,<0.4
Pillow>=10.0,<12.0
PyPDF2>=3.0.0,<4.0
